Added new libraries and sky box system.
* Added two new libraries: tinygltf and Optik.
* Flint can now load skyboxes using HDR images (thanks to Nadrin).
* Flint now has mip chain support.

+++ Code/GraphicsCore/Source/GeometryStore.cpp
@@ -213,8 +213,11 @@ namespace Flint
 
 	void GeometryStore::Terminate()
 	{
-		pVertexBuffer->Terminate();
-		pIndexBuffer->Terminate();
+		if (pVertexBuffer)
+			pVertexBuffer->Terminate();
+
+		if (pIndexBuffer)
+			pIndexBuffer->Terminate();
 
 		bIsTerminated = true;
 	}
